Refactor ESP32 transport modules and sync docs
Agent-Logs-Url: https://github.com/Amperstrand/microfips/sessions/5f1c0557-527d-4d3c-8981-6461782b023b
