+++ crates/microfips-esp32/src/handler.rs
@@ -1,0 +1,85 @@
+use core::sync::atomic::Ordering;
+
+use microfips_http_demo::DemoService;
+use microfips_protocol::fsp_handler::FspDualHandler;
+use microfips_protocol::node::{HandleResult, NodeEvent, NodeHandler};
+use microfips_service::FspServiceAdapter;
+
+use crate::config::{ESP32_SECRET, LED_OFF, LED_ON, STM32_NODE_ADDR, STM32_PEER_PUB};
+use crate::led::Led;
+use crate::stats::{
+    STAT_DATA_RX, STAT_DATA_TX, STAT_HB_RX, STAT_HB_TX, STAT_MSG1_TX, STAT_MSG2_RX,
+};
+
+pub type EspFspHandler = FspDualHandler<FspServiceAdapter<DemoService>>;
+
+pub fn build_demo_fsp(
+    responder_ephemeral: [u8; 32],
+    initiator_ephemeral: [u8; 32],
+) -> EspFspHandler {
+    FspDualHandler::new_dual(
+        ESP32_SECRET,
+        responder_ephemeral,
+        initiator_ephemeral,
+        &STM32_PEER_PUB,
+        STM32_NODE_ADDR,
+        FspServiceAdapter::new(DemoService::new()),
+    )
+}
+
+pub struct EspHandler<'a> {
+    pub led: &'a mut Led,
+    pub fsp: EspFspHandler,
+}
+
+impl NodeHandler for EspHandler<'_> {
+    async fn on_event(&mut self, event: NodeEvent) {
+        match event {
+            NodeEvent::Connected => {
+                self.led.set_state(LED_ON);
+            }
+            NodeEvent::Msg1Sent => {
+                STAT_MSG1_TX.fetch_add(1, Ordering::Relaxed);
+                self.led.set_state(LED_ON);
+            }
+            NodeEvent::HandshakeOk => {
+                STAT_MSG2_RX.fetch_add(1, Ordering::Relaxed);
+                self.led.set_state(LED_ON);
+            }
+            NodeEvent::HeartbeatSent => {
+                STAT_HB_TX.fetch_add(1, Ordering::Relaxed);
+            }
+            NodeEvent::HeartbeatRecv => {
+                STAT_HB_RX.fetch_add(1, Ordering::Relaxed);
+            }
+            NodeEvent::Disconnected | NodeEvent::Error => {
+                self.led.set_state(LED_OFF);
+            }
+        }
+        self.fsp.on_event_default(event);
+    }
+
+    fn on_message(&mut self, msg_type: u8, payload: &[u8], resp: &mut [u8]) -> HandleResult {
+        if msg_type != 0x00 {
+            return HandleResult::None;
+        }
+        STAT_DATA_RX.fetch_add(1, Ordering::Relaxed);
+        let result = self.fsp.on_message(msg_type, payload, resp);
+        if let HandleResult::SendDatagram(_) = result {
+            STAT_DATA_TX.fetch_add(1, Ordering::Relaxed);
+        }
+        result
+    }
+
+    fn poll_at(&self) -> Option<embassy_time::Instant> {
+        self.fsp.poll_at()
+    }
+
+    fn on_tick(&mut self, resp: &mut [u8]) -> HandleResult {
+        let result = self.fsp.on_tick(resp);
+        if let HandleResult::SendDatagram(_) = result {
+            STAT_DATA_TX.fetch_add(1, Ordering::Relaxed);
+        }
+        result
+    }
+}
